refactor: features/tag 제거
