+++ frontend/techpick/src/components/DeleteTagDialog/index.tsx
@@ -7,7 +7,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import { notifyError } from '@/utils';
 import { Text, Button, Gap } from '@/components';
 import { useTagStore } from '@/stores/tagStore';
-import { useDeleteTagDialogStore } from '../../deleteTag.model';
+import { useDeleteTagDialogStore } from '@/stores/deleteTagDialogStore';
 import { dialogContentStyle, dialogOverlayStyle } from './DeleteTagDialog.css';
 
 export const DeleteTagDialog = memo(function DeleteTagDialog() {
