+++ frontend/techpick/src/components/DeselectTagButton/DeselectTagButton.tsx
