--- frontend/techpick/src/features/tag/deleteTag/index.ts
@@ -1,2 +1,0 @@
-export { ShowDeleteTagDialogButton } from './ui/ShowDeleteTagDialogButton';
-export * from './ui/DeleteTagDialog';
