--- frontend/techpick/src/features/tag/index.ts
@@ -1,3 +1,0 @@
-export * from './deselectTag/DeselectTagButton';
-export * from './deleteTag';
-export * from './deleteTag/deleteTag.model';
